section api added

+++ Nspiration/Model/Folder.cs
@@ -9,8 +9,8 @@ namespace Nspiration.Model
         public string? Name { get; set; }
         public bool IsActive { get; set; }
         public DateTime CreatedAt { get; set; }
-        public long CreatedBy { get; set; }
+        public int CreatedBy { get; set; }
         public DateTime? ModifiedAt { get; set; }
-        public long? ModifiedBy { get; set; }
+        public int? ModifiedBy { get; set; }
     }
 }
